% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_T_half_1cpt.R
\name{alpha.T12}
\alias{alpha.T12}
\title{Calculate half-life from 1-compartmental models}
\usage{
alpha.T12(V, Cl, kel)
}
\arguments{
\item{V}{Volume of distribution for the central compartment.}

\item{Cl}{Clearance from the central compartment.}

\item{kel}{Elimination rate constant from central compartment.}
}
\value{
The calculated \eqn{\alpha} half-life as a single value
}
\description{
This function will calculate the \eqn{\alpha} half-life for 1-cpt models. The time units for the parameters given will be used for the units of the half-life.
The user needs to provide the \code{V} with either \code{kel} or \code{Cl}. If both \code{kel} and \code{Cl} are provided, the kel will be used for half-life calculation.
}
\examples{
alpha.T12(V   = 3,    # in L
          Cl  = 16)   # in L/day
Output: [1] 0.1299651  # in days
alpha.T12(V   = 3     # in L
          kel = 5.33) # in 1/day
Output: [1] 0.1300464  # in days

}
